+++ Assets/Scripts/UI/NewMapMenu.cs
@@ -1,9 +1,10 @@
 ﻿using UnityEngine;
+using UnityEngine.UIElements;
 
 /// <summary>
 /// Component that applies actions from the new map menu UI to the hex map.
-/// Public methods are hooked up to the in-game UI.
 /// </summary>
+[RequireComponent(typeof(UIDocument))]
 public class NewMapMenu : MonoBehaviour
 {
 	[SerializeField]
@@ -16,14 +17,27 @@ public class NewMapMenu : MonoBehaviour
 
 	bool wrapping = true;
 
-	public void ToggleMapGeneration(bool toggle) => generateMaps = toggle;
-
-	public void ToggleWrapping(bool toggle) => wrapping = toggle;
-
 	public void Open()
 	{
 		gameObject.SetActive(true);
 		HexMapCamera.Locked = true;
+
+		VisualElement root = GetComponent<UIDocument>().rootVisualElement;
+		
+		var generateToggle = root.Q<Toggle>("Generate");
+		generateToggle.value = generateMaps;
+		generateToggle.RegisterValueChangedCallback(
+			change => generateMaps = change.newValue);
+
+		var wrappingToggle = root.Q<Toggle>("Wrapping");
+		wrappingToggle.value = wrapping;
+		wrappingToggle.RegisterValueChangedCallback(
+			change => wrapping = change.newValue);
+		
+		root.Q<Button>("Small").clicked += () => CreateMap(20, 15);
+		root.Q<Button>("Medium").clicked += () => CreateMap(40, 30);
+		root.Q<Button>("Large").clicked += () => CreateMap(80, 60);
+		root.Q<Button>("Cancel").clicked += Close;
 	}
 
 	public void Close()
@@ -31,12 +45,6 @@ public class NewMapMenu : MonoBehaviour
 		gameObject.SetActive(false);
 		HexMapCamera.Locked = false;
 	}
-
-	public void CreateSmallMap() => CreateMap(20, 15);
-
-	public void CreateMediumMap() => CreateMap(40, 30);
-
-	public void CreateLargeMap() => CreateMap(80, 60);
 
 	void CreateMap(int x, int z)
 	{