+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Upgraded to Unity 6000.0.43f1, Burst 1.8.19, and URP 17.0.4.
+
+### Fixed
+- Check correct levels when editing farm or plant levels.
+
 ## [3.4.0] - 2024-07-30
 
 ### Changed