+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.2.0] - 2025-02-26
+
+### Changed
+
+- Migrated Save Load menu from uGUI to UI Toolkit.
+
+## [5.1.0] - 2025-02-20
+
+### Changed
+
+- Upgraded Input System to 1.18.0.
+- Migrated New Map menu from uGUI to UI Toolkit.
+
+## [5.0.0] - 2025-01-19
+
+### Changed
+
+- Upgraded Unity to 6000.3.1f1, Burst 1.8.27, and URP 17.3.0.
+- Switched to using the new Input System package, version 1.17.0.
+
 ## [4.1.0] - 2025-08-28
 
 ### Added